EpicLoot 0.5.9
Fixed bug where sacrificing with nearly full inventory resulted in lost items (items that do not fit in the inventory now fall to the ground)
Loot2 and Loot3 in the loot table are now exclusive (e.g. if the mob is level 2, and Loot2 is present, then only the loot from Loot2 is used, otherwise it falls back to Loot)
ItemSets have been added to the loot table schema (If the "Item" field of the Loot list is in the item set, roll on that loot table for that item instead)
Can now reference other loot tables in the loot table item config using "Item": "<object>.<level>" where <object> is the name of a loot table entry and <level> is an integer between 1 and 3 that refers to Loot, Loot2, or Loot3.
Number of magic effects per rarity is now configurable in loottables.json in the "MagicEffectsCount" object
Completed loottables with updates from feedback and using the new system