Update TextsDialog_Patch.cs
This code will destroy all child objects of the TextContainer. If m_textArea is one of the child objects of the TextContainer and is being used when the child objects are being destroyed, it will cause m_textArea to become null. This can result in other mods being unable to access the m_textArea object.